Adding doc in new index
