+++ components/compliance-service/ingest/pipeline/processor/control_worker.go
@@ -2,12 +2,10 @@ package processor
 
 import (
 	"context"
-	"fmt"
 	"github.com/chef/automate/components/compliance-service/ingest/ingestic"
 	"github.com/chef/automate/components/compliance-service/ingest/ingestic/mappings"
 	"github.com/chef/automate/components/compliance-service/reporting/relaxting"
 	"github.com/chef/automate/lib/cereal"
-	elastic "github.com/olivere/elastic/v7"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"time"
@@ -169,8 +167,6 @@ func (t *GenerateControlTask) Run(ctx context.Context, task cereal.Task) (interf
 
 	logrus.Infof("In TaskRun working on job %s", job.ReportUuid)
 
-	//startTime := time.Now().UTC().Round(time.Second)
-
 	//time taking to commit to records to ES
 	time.Sleep(60 * time.Second)
 	mapping := mappings.ComplianceRepDate
@@ -182,95 +178,11 @@ func (t *GenerateControlTask) Run(ctx context.Context, task cereal.Task) (interf
 		return nil, err
 	}
 
-	err = UploadDataToControlIndex(ctx, job.ReportUuid, controls, t.ESClient, job.EndTime)
+	err = t.ESClient.UploadDataToControlIndex(ctx, job.ReportUuid, controls, job.EndTime)
+	if err != nil {
+		logrus.Errorf("Unable to add data to index with reportuuid:%s", job.ReportUuid)
+		return nil, err
+	}
 
 	return &job, nil
-}
-
-func UploadDataToControlIndex(ctx context.Context, reportuuid string, controls []Control, client *ingestic.ESClient, endTime time.Time) error {
-	mapping := mappings.ComplianceControlRepData
-	index := mapping.IndexTimeseriesFmt(endTime)
-
-	bulkRequest := client.Client.Bulk()
-	for _, control := range controls {
-		docId := GetDocIdByControlIdAndProfileID(control.ControlID, control.Profile.ProfileID)
-		found, err := CheckIfControlIdExistsForToday(docId, index, client)
-		if err != nil {
-			logrus.Errorf("Unable to fetch document for control id %s|%s", control.ControlID, control.Profile.ProfileID)
-		}
-		if found {
-			bulkRequest = bulkRequest.Add(elastic.NewBulkUpdateRequest().Index(index).Id(docId).Script(createScriptForAddingNode(control.Nodes[0])).Type("_doc"))
-		} else {
-			bulkRequest = bulkRequest.Add(elastic.NewBulkIndexRequest().Index(index).Id(docId).Doc(control).Type("_doc"))
-		}
-
-	}
-
-	approxBytes := bulkRequest.EstimatedSizeInBytes()
-	bulkResponse, err := bulkRequest.Refresh("false").Do(ctx)
-	if err != nil {
-		logrus.Errorf("Unable to send the request in bulk for reportuuid :%s with error :%v", reportuuid, err)
-		return err
-	}
-	if bulkResponse == nil {
-		logrus.Errorf("Unable to fetch the response of bulk request reportuuid id:%s with error :%v", reportuuid, err)
-		return err
-	}
-
-	logrus.Debugf("Bulk insert %d summaries, ~size %dB, took %dms", len(controls), approxBytes, bulkResponse.Took)
-	return nil
-
-}
-
-func GetDocIdByControlIdAndProfileID(controlID string, profileID string) string {
-	return fmt.Sprintf("%s|%s", controlID, profileID)
-}
-
-func createScriptForAddingNode(node Node) *elastic.Script {
-	params := make(map[string]interface{})
-	params["node"] = node
-
-	return elastic.NewScript("if (!(ctx._source.nodes instanceof Collection)) {ctx._source.nodes = [ctx._source.nodes];} ctx._source.nodes.add(params.node)").Params(params)
-
-}
-
-func CheckIfControlIdExistsForToday(docId string, indexToday string, esClient *ingestic.ESClient) (bool, error) {
-	logrus.Debugf("Checking the control document exists for today with doc Id :%s", docId)
-	fsc := elastic.NewFetchSourceContext(false)
-	boolQuery := elastic.NewBoolQuery()
-	idsQuery := elastic.NewIdsQuery()
-	idsQuery.Ids(docId)
-	boolQuery = boolQuery.Must(idsQuery)
-	searchSource := elastic.NewSearchSource().
-		FetchSourceContext(fsc).
-		Query(boolQuery).
-		Size(1)
-	searchResult, err := esClient.Client.Search().
-		SearchSource(searchSource).
-		Index(indexToday).
-		Do(context.Background())
-	if err != nil {
-		switch {
-		case elastic.IsTimeout(err):
-			logrus.Errorf("Timeout retrieving document: %v", err)
-			return false, err
-		default:
-			logrus.Errorf("Received error: %v", err)
-			return false, err
-		}
-	}
-
-	if searchResult.TotalHits() > 0 {
-		// Iterate through results
-		for _, hit := range searchResult.Hits.Hits {
-			// hit.Index contains the id of the index
-			if len(hit.Id) > 0 {
-				logrus.Debugf("Found the document with for control with doc Id %s", docId)
-				return true, nil
-			}
-
-		}
-	}
-	return false, nil
-
 }
